+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpLogExporterTests.cs
@@ -21,6 +21,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Moq;
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
@@ -592,9 +593,10 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
         var emptyLogRecords = Array.Empty<LogRecord>();
         var emptyBatch = new Batch<LogRecord>(emptyLogRecords, emptyLogRecords.Length);
         var sut = new OtlpLogExporter(
-                        new OtlpExporterOptions(),
-                        new SdkLimitOptions(),
-                        fakeExportClient.Object);
+            new OtlpExporterOptions(),
+            new SdkLimitOptions(),
+            new ExperimentalOptions(),
+            fakeExportClient.Object);
 
         // Act.
         var result = sut.Export(emptyBatch);
@@ -614,9 +616,10 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
             .Setup(_ => _.SendExportRequest(It.IsAny<OtlpCollector.ExportLogsServiceRequest>(), default))
             .Throws(new Exception("Test Exception"));
         var sut = new OtlpLogExporter(
-                        new OtlpExporterOptions(),
-                        new SdkLimitOptions(),
-                        fakeExportClient.Object);
+            new OtlpExporterOptions(),
+            new SdkLimitOptions(),
+            new ExperimentalOptions(),
+            fakeExportClient.Object);
 
         // Act.
         var result = sut.Export(emptyBatch);
@@ -636,9 +639,10 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
             .Setup(_ => _.SendExportRequest(It.IsAny<OtlpCollector.ExportLogsServiceRequest>(), default))
             .Returns(true);
         var sut = new OtlpLogExporter(
-                        new OtlpExporterOptions(),
-                        new SdkLimitOptions(),
-                        fakeExportClient.Object);
+            new OtlpExporterOptions(),
+            new SdkLimitOptions(),
+            new ExperimentalOptions(),
+            fakeExportClient.Object);
 
         // Act.
         var result = sut.Export(emptyBatch);
@@ -1175,11 +1179,13 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
 
         options.AddOtlpExporter();
 
-        var processors = (List<BaseProcessor<LogRecord>>)typeof(OpenTelemetryLoggerOptions).GetField("Processors", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(options);
+        var provider = new OpenTelemetryLoggerProvider(new TestOptionsMonitor<OpenTelemetryLoggerOptions>(options));
 
-        Assert.Single(processors);
+        var processor = GetProcessor(provider);
 
-        var batchProcesor = processors[0] as BatchLogRecordExportProcessor;
+        Assert.NotNull(processor);
+
+        var batchProcesor = processor as BatchLogRecordExportProcessor;
 
         Assert.NotNull(batchProcesor);
 
@@ -1201,13 +1207,15 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
             l.BatchExportProcessorOptions = new BatchExportLogRecordProcessorOptions() { ScheduledDelayMilliseconds = 1000 };
         });
 
-        var processors = (List<BaseProcessor<LogRecord>>)typeof(OpenTelemetryLoggerOptions).GetField("Processors", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(options);
+        var provider = new OpenTelemetryLoggerProvider(new TestOptionsMonitor<OpenTelemetryLoggerOptions>(options));
 
-        Assert.Single(processors);
+        var processor = GetProcessor(provider);
+
+        Assert.NotNull(processor);
 
         if (processorType == ExportProcessorType.Batch)
         {
-            var batchProcesor = processors[0] as BatchLogRecordExportProcessor;
+            var batchProcesor = processor as BatchLogRecordExportProcessor;
 
             Assert.NotNull(batchProcesor);
 
@@ -1217,7 +1225,7 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
         }
         else
         {
-            var simpleProcesor = processors[0] as SimpleLogRecordExportProcessor;
+            var simpleProcesor = processor as SimpleLogRecordExportProcessor;
 
             Assert.NotNull(simpleProcesor);
         }
@@ -1281,8 +1289,153 @@ public class OtlpLogExporterTests : Http2UnencryptedSupportTests
         Assert.Empty(OtlpLogRecordTransformer.LogListPool);
     }
 
+    [Theory]
+    [InlineData(null)]
+    [InlineData("logging")]
+    public void VerifyEnvironmentVariablesTakenFromIConfigurationWhenUsingLoggerFactoryCreate(string optionsName)
+    {
+        RunVerifyEnvironmentVariablesTakenFromIConfigurationTest(
+            optionsName,
+            configure =>
+            {
+                var factory = LoggerFactory.Create(logging =>
+                {
+                    configure(logging.Services);
+
+                    logging.AddOpenTelemetry(o => o.AddOtlpExporter(optionsName, configure: null));
+                });
+
+                return (factory, factory);
+            });
+    }
+
+    [Theory]
+    [InlineData(null)]
+    [InlineData("logging")]
+    public void VerifyEnvironmentVariablesTakenFromIConfigurationWhenUsingLoggingBuilder(string optionsName)
+    {
+        RunVerifyEnvironmentVariablesTakenFromIConfigurationTest(
+            optionsName,
+            configure =>
+            {
+                var services = new ServiceCollection();
+
+                configure(services);
+
+                services.AddLogging(
+                    logging => logging.AddOpenTelemetry(o =>
+                        o.AddOtlpExporter(optionsName, configure: null)));
+
+                var sp = services.BuildServiceProvider();
+
+                var factory = sp.GetRequiredService<ILoggerFactory>();
+
+                return (sp, factory);
+            });
+    }
+
+    private static void RunVerifyEnvironmentVariablesTakenFromIConfigurationTest(
+        string optionsName,
+        Func<Action<IServiceCollection>, (IDisposable Container, ILoggerFactory LoggerFactory)> createLoggerFactoryFunc)
+    {
+        var values = new Dictionary<string, string>()
+        {
+            [OtlpExporterOptions.EndpointEnvVarName] = "http://test:8888",
+        };
+
+        var configuration = new ConfigurationBuilder()
+            .AddInMemoryCollection(values)
+            .Build();
+
+        var configureDelegateCalled = false;
+        var configureExportProcessorOptionsCalled = false;
+        var configureBatchOptionsCalled = false;
+
+        var tracingConfigureDelegateCalled = false;
+        var unnamedConfigureDelegateCalled = false;
+        var allConfigureDelegateCalled = false;
+
+        var testState = createLoggerFactoryFunc(services =>
+        {
+            services.AddSingleton<IConfiguration>(configuration);
+
+            services.Configure<OtlpExporterOptions>(optionsName, o =>
+            {
+                configureDelegateCalled = true;
+                Assert.Equal(new Uri("http://test:8888"), o.Endpoint);
+            });
+
+            services.Configure<LogRecordExportProcessorOptions>(optionsName, o =>
+            {
+                configureExportProcessorOptionsCalled = true;
+            });
+
+            services.Configure<BatchExportLogRecordProcessorOptions>(optionsName, o =>
+            {
+                configureBatchOptionsCalled = true;
+            });
+
+            services.Configure<OtlpExporterOptions>("tracing", o =>
+            {
+                tracingConfigureDelegateCalled = true;
+            });
+
+            services.Configure<OtlpExporterOptions>(o =>
+            {
+                unnamedConfigureDelegateCalled = true;
+            });
+
+            services.ConfigureAll<OtlpExporterOptions>(o =>
+            {
+                allConfigureDelegateCalled = true;
+            });
+        });
+
+        using var container = testState.Container;
+
+        var factory = testState.LoggerFactory;
+
+        Assert.NotNull(factory);
+
+        Assert.True(configureDelegateCalled);
+        Assert.True(configureExportProcessorOptionsCalled);
+        Assert.True(configureBatchOptionsCalled);
+
+        Assert.False(tracingConfigureDelegateCalled);
+
+        Assert.Equal(optionsName == null, unnamedConfigureDelegateCalled);
+
+        Assert.True(allConfigureDelegateCalled);
+    }
+
     private static OtlpCommon.KeyValue TryGetAttribute(OtlpLogs.LogRecord record, string key)
     {
         return record.Attributes.FirstOrDefault(att => att.Key == key);
+    }
+
+    private static BaseProcessor<LogRecord> GetProcessor(OpenTelemetryLoggerProvider provider)
+    {
+        var sdkProvider = typeof(OpenTelemetryLoggerProvider).GetField("Provider", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(provider);
+
+        return (BaseProcessor<LogRecord>)sdkProvider.GetType().GetProperty("Processor", BindingFlags.Instance | BindingFlags.Public).GetMethod.Invoke(sdkProvider, null);
+    }
+
+    private sealed class TestOptionsMonitor<T> : IOptionsMonitor<T>
+    {
+        private readonly T instance;
+
+        public TestOptionsMonitor(T instance)
+        {
+            this.instance = instance;
+        }
+
+        public T CurrentValue => this.instance;
+
+        public T Get(string name) => this.instance;
+
+        public IDisposable OnChange(Action<T, string> listener)
+        {
+            throw new NotImplementedException();
+        }
     }
 }